requirements.txt
streamlit
pandas
joblib
shap
matplotlib
scikit-learn
Pillow
